pyproject.toml
[tool.poetry]
name = "callie-integrations"
version = "1.0.0"
description = "Enterprise data synchronization platform for Calibrate Network"
authors = ["Calibrate Network <dev@calibratenetwork.com>"]
readme = "README.md"
packages = [{include = "callie", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.32.4"
pydantic = "^2.10.6"
click = "^8.1.8"
python-dotenv = "^1.0.1"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
google-cloud-firestore = "^2.20.0"
google-cloud-scheduler = "^2.14.0"
google-auth = "^2.36.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.0.0"
flake8 = "^6.0.0"
mypy = "^1.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
callie = "callie.core.cli:main"
callie-api = "callie.api.app:main" 